Latest commit in this branch

+++ blog-app/src/app/components/other/dashboard/dashboard.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { AccountService } from '../../../service/account.service';
 import { CommonModule } from '@angular/common';
+import { Writer } from '../../../interface/isWriter.user';
+import { Admin } from '../../../interface/isAdmin.user';
 
 @Component({
   selector: 'app-dashboard',
@@ -15,15 +17,31 @@ export class DashboardComponent {
 
   constructor(private accountService: AccountService){ }
 
+  userName: string = ''
   userEmail: string = ''
-  deletingUserError = this.accountService.deletingUserError
+  userStatus: string = ''
+  deletingUserError: boolean = false
 
   ngOnInit() {
     if(typeof(Storage) !== 'undefined'){
       const userData = localStorage.getItem('user')
       if(userData) {
+
+        // Check if user is administrator or writer
         let user = JSON.parse(userData)
         this.userEmail = user.email
+        const adminInArray = Admin.find(u => u.email === user.email) // Find the admin in the array
+        const writterInArray = Writer.find(u => u.email === user.email)
+
+        if(adminInArray && adminInArray.isAdmin === true) {
+          this.userStatus = 'Administrator'
+        } else if(writterInArray && writterInArray.isWriter === true) {
+          this.userStatus = 'Pisarz'
+        } else {
+          this.userStatus = 'Użytkownik'
+        }
+
+        // Create user data in database
       }
     }
   }
@@ -32,7 +50,12 @@ export class DashboardComponent {
     this.accountService.changePassword()
   }
   deleteUserAccount() {
-    this.accountService.deleteUserAccount()
+    if (this.accountService.deletingUserError){
+      this.deletingUserError = true
+    } else {
+      this.accountService.deleteUserAccount()
+    }
+
   }
